Change Payment Cost type to double

+++ src/ShopOnline/Model/Payment.cs
@@ -5,13 +5,13 @@ namespace ShopOnline.Model
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Cost { get; set; }
+        public double Cost { get; set; }
 
         public Payment()
         {
         }
 
-        public Payment(int id, string name, int cost)
+        public Payment(int id, string name, double cost)
         {
             Id = id;
             Name = name;
